requirements.txt
click==8.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dotenv==1.1.0
rich==14.0.0
shellingham==1.5.4
SQLAlchemy==2.0.41
typer==0.16.0
typing_extensions==4.13.2